7 - move interceptor to apiConfig
