+++ src/domain/Auth/authApi.ts
@@ -75,6 +75,10 @@ async function refreshToken(token: string): Promise<AuthCredentialsAPI> {
   return response.data;
 }
 
+/**
+ * @param axiosConfig [AxiosRequestConfig](https://axios-http.com/docs/req_config) - The Axios request configuration
+ * @returns  Check the config URL property to returns if is a refresh token request
+ */
 function isRefreshTokenRequest(request: AxiosRequestConfig): boolean {
   const url = request.url;
   return url === REFRESH_TOKEN_URL;
