+++ src/services/authCredentials/Providers/AuthCredentialsProvider.tsx
@@ -1,10 +1,9 @@
 import React, {useEffect} from 'react';
 import {createContext, useState} from 'react';
 
-import {api} from '@api';
+import {registerInterceptor} from '@api';
 import {AuthCredentials, authService} from '@domain';
 
-import {authApi} from '../../../domain/Auth/authApi';
 import {authCredentialsStorage} from '../authCredentialsStorage';
 import {AuthCredentialsService} from '../authCredentialsTypes';
 
@@ -28,42 +27,15 @@ export function AuthCredentialsProvider({
   }, []);
 
   useEffect(() => {
-    const interceptor = api.interceptors.response.use(
-      response => response,
-      async responseError => {
-        const failedRequest = responseError.config;
-        const hasNotRefreshToken = !authCredentials?.refreshToken;
-        const isRefreshTokenRequest =
-          authApi.isRefreshTokenRequest(failedRequest);
-
-        if (responseError.response.status === 401) {
-          if (
-            hasNotRefreshToken ||
-            isRefreshTokenRequest ||
-            failedRequest.sent
-          ) {
-            removeCredentials();
-            return Promise.reject(responseError);
-          }
-
-          failedRequest.sent = true;
-
-          const newAuthCredentials =
-            await authService.authenticateByRefreshToken(
-              authCredentials?.refreshToken,
-            );
-          saveCredentials(newAuthCredentials);
-
-          failedRequest.headers.Authorization = `Bearer ${newAuthCredentials.token}`;
-
-          return api(failedRequest);
-        }
-      },
-    );
+    const interceptor = registerInterceptor({
+      authCredentials,
+      removeCredentials,
+      saveCredentials,
+    });
 
     // remove listener when component unmount
-    return () => api.interceptors.response.eject(interceptor);
-  }, [authCredentials?.refreshToken]);
+    return interceptor;
+  }, [authCredentials]);
 
   async function startAuthCredentials() {
     try {
